feat: refactor project structure and implemented user endpoint
