+++ src/modules/users/controllers/users.controller.ts
@@ -4,19 +4,22 @@ import {
   Post,
   Body,
   Param,
-  Patch,
   Delete,
+  HttpStatus,
+  HttpCode,
+  Put,
 } from '@nestjs/common';
-import { UsersService } from './users.service';
-import { CreateUserDto } from './dto/create-user.dto';
-import { UpdateUserDto } from './dto/update-user.dto';
-import { User } from './user.interface';
+import { UsersService } from '../services/users.service';
+import { CreateUserDto } from './../dto/create-user.dto';
+import { UpdateUserDto } from './../dto/update-user.dto';
+import { User } from './../interfaces/user.interface';
 
 @Controller('users')
 export class UsersController {
   constructor(private readonly usersService: UsersService) {}
 
   @Post()
+  @HttpCode(HttpStatus.CREATED)
   create(@Body() createUserDto: CreateUserDto): User {
     return this.usersService.create(createUserDto);
   }
@@ -31,12 +34,13 @@ export class UsersController {
     return this.usersService.findOne(id);
   }
 
-  @Patch(':id')
+  @Put(':id')
   update(@Param('id') id: string, @Body() updateUserDto: UpdateUserDto): User {
     return this.usersService.update(id, updateUserDto);
   }
 
   @Delete(':id')
+  @HttpCode(HttpStatus.NO_CONTENT)
   remove(@Param('id') id: string): void {
     this.usersService.remove(id);
   }
