--- src/users/dto/create-user.dto.ts
@@ -1,11 +1,0 @@
-import { IsString, Length } from 'class-validator';
-
-export class CreateUserDto {
-  @IsString()
-  @Length(3, 20)
-  login: string;
-
-  @IsString()
-  @Length(8, 100)
-  password: string;
-}
