--- src/users/dto/update-user.dto.ts
@@ -1,8 +1,0 @@
-import { PartialType } from '@nestjs/mapped-types';
-import { IsNumber } from 'class-validator';
-import { CreateUserDto } from './create-user.dto';
-
-export class UpdateUserDto extends PartialType(CreateUserDto) {
-  @IsNumber()
-  version: number;
-}
