--- src/users/users.module.ts
@@ -1,9 +1,0 @@
-import { Module } from '@nestjs/common';
-import { UsersController } from './users.controller';
-import { UsersService } from './users.service';
-
-@Module({
-  controllers: [UsersController],
-  providers: [UsersService]
-})
-export class UsersModule {}
