--- src/users/users.service.ts
@@ -1,45 +1,0 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
-import { User } from './user.interface';
-import { CreateUserDto } from './dto/create-user.dto';
-import { randomUUID } from 'crypto';
-import { UpdateUserDto } from './dto/update-user.dto';
-
-@Injectable()
-export class UsersService {
-  private users: User[] = [];
-
-  create(createUserDto: CreateUserDto): User {
-    const newUser: User = {
-      id: randomUUID(),
-      login: createUserDto.login,
-      password: createUserDto.password,
-      version: 1,
-      createdAt: Date.now(),
-      updatedAt: Date.now(),
-    };
-    this.users.push(newUser);
-    return newUser;
-  }
-
-  findAll(): User[] {
-    return this.users;
-  }
-
-  findOne(id: string): User {
-    const user = this.users.find((item) => item.id === id);
-    if (!user) throw new NotFoundException('User not found');
-    return user;
-  }
-
-  update(id: string, updateUserDto: UpdateUserDto): User {
-    const user = this.findOne(id);
-    Object.assign(user, updateUserDto);
-    return user;
-  }
-
-  remove(id: string) {
-    const index = this.users.findIndex((item) => item.id === id);
-    if (index === -1) throw new NotFoundException('User not found');
-    this.users.splice(index, 1);
-  }
-}
